Refactored Scene; Added TransformComponent; Refactor Submit() of Renderer and PhysicsSimulator; Changed `ShouldRenderInViewport` to `Display` in EntityInfo; Added `Internal` in EntityInfo; Added extra message before `Save As` message when creating or opening a project while the current project is unsaved;

+++ runtime/src/function/ui/imgui_wrapper.cpp
@@ -16,13 +16,13 @@ ARCADIA_API auto Arcadia::ImguiWrapper::Checkbox(const std::string& name, std::f
     return deactivated;
 }
 
-ARCADIA_API auto Arcadia::ImguiWrapper::DragInt(const std::string& name, int& i, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool
+ARCADIA_API auto Arcadia::ImguiWrapper::DragInt(const std::string& name, int& i, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool
 {
     ImGui::Text(name.c_str()); ImGui::SameLine(); ImGui::DragInt(std::format("##{}", name).c_str(), &i, speed, min, max, format, flags);
     return ImGui::IsItemDeactivatedAfterEdit();
 }
 
-ARCADIA_API auto Arcadia::ImguiWrapper::DragInt(const std::string& name, std::function<int()> getter, std::function<void(int)> setter, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool
+ARCADIA_API auto Arcadia::ImguiWrapper::DragInt(const std::string& name, std::function<int()> getter, std::function<void(int)> setter, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool
 {
     auto i = getter();
     auto deactivated = Arcadia::ImguiWrapper::DragInt(name, i, speed, min, max, format, flags);
@@ -30,7 +30,7 @@ ARCADIA_API auto Arcadia::ImguiWrapper::DragInt(const std::string& name, std::fu
     return deactivated;
 }
 
-ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec2(const std::string& name, glm::ivec2& vec, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool
+ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec2(const std::string& name, glm::ivec2& vec, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool
 {
     bool deactivated{ false };
     ImGui::Text(std::format("{:>{}} X", name, name.size()).c_str()); ImGui::SameLine(); ImGui::DragInt(std::format("##{}_x", name).c_str(), &vec.x, speed, min, max, format, flags);
@@ -46,7 +46,7 @@ ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec2(const std::string& name, glm::
     return deactivated;
 }
 
-ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec2(const std::string& name, std::function<glm::ivec2()> getter, std::function<void(const glm::ivec2&)> setter, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool
+ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec2(const std::string& name, std::function<glm::ivec2()> getter, std::function<void(const glm::ivec2&)> setter, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool
 {
     auto vec = getter();
     auto deactivated = Arcadia::ImguiWrapper::DragIVec2(name, vec, speed, min, max, format, flags);
@@ -54,7 +54,7 @@ ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec2(const std::string& name, std::
     return deactivated;
 }
 
-ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec3(const std::string& name, glm::ivec3& vec, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool
+ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec3(const std::string& name, glm::ivec3& vec, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool
 {
     bool deactivated{ false };
     ImGui::Text(std::format("{:>{}} X", name, name.size()).c_str()); ImGui::SameLine(); ImGui::DragInt(std::format("##{}_x", name).c_str(), &vec.x, speed, min, max, format, flags);
@@ -75,7 +75,7 @@ ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec3(const std::string& name, glm::
     return deactivated;
 }
 
-ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec3(const std::string& name, std::function<glm::ivec3()> getter, std::function<void(const glm::ivec3&)> setter, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool
+ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec3(const std::string& name, std::function<glm::ivec3()> getter, std::function<void(const glm::ivec3&)> setter, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool
 {
     auto vec = getter();
     auto deactivated = Arcadia::ImguiWrapper::DragIVec3(name, vec, speed, min, max, format, flags);
@@ -83,7 +83,7 @@ ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec3(const std::string& name, std::
     return deactivated;
 }
 
-ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec4(const std::string& name, glm::ivec4& vec, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool
+ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec4(const std::string& name, glm::ivec4& vec, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool
 {
     bool deactivated{ false };
     ImGui::Text(std::format("{:>{}} X", name, name.size()).c_str()); ImGui::SameLine(); ImGui::DragInt(std::format("##{}_x", name).c_str(), &vec.x, speed, min, max, format, flags);
@@ -109,7 +109,7 @@ ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec4(const std::string& name, glm::
     return deactivated;
 }
 
-ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec4(const std::string& name, std::function<glm::ivec4()> getter, std::function<void(const glm::ivec4&)> setter, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool
+ARCADIA_API auto Arcadia::ImguiWrapper::DragIVec4(const std::string& name, std::function<glm::ivec4()> getter, std::function<void(const glm::ivec4&)> setter, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool
 {
     auto vec = getter();
     auto deactivated = Arcadia::ImguiWrapper::DragIVec4(name, vec, speed, min, max, format, flags);
