+++ runtime/src/function/ui/imgui_wrapper.hpp
@@ -13,17 +13,17 @@ namespace Arcadia
         ARCADIA_API auto Checkbox(const std::string& name, bool& b) -> bool;
         ARCADIA_API auto Checkbox(const std::string& name, std::function<bool()> getter, std::function<void(bool)> setter) -> bool;
 
-        ARCADIA_API auto DragInt(const std::string& name, int& i, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool;
-        ARCADIA_API auto DragInt(const std::string& name, std::function<int()> getter, std::function<void(int)> setter, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool;
+        ARCADIA_API auto DragInt(const std::string& name, int& i, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool;
+        ARCADIA_API auto DragInt(const std::string& name, std::function<int()> getter, std::function<void(int)> setter, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool;
 
-        ARCADIA_API auto DragIVec2(const std::string& name, glm::ivec2& vec, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool;
-        ARCADIA_API auto DragIVec2(const std::string& name, std::function<glm::ivec2()> getter, std::function<void(const glm::ivec2&)> setter, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool;
+        ARCADIA_API auto DragIVec2(const std::string& name, glm::ivec2& vec, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool;
+        ARCADIA_API auto DragIVec2(const std::string& name, std::function<glm::ivec2()> getter, std::function<void(const glm::ivec2&)> setter, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool;
 
-        ARCADIA_API auto DragIVec3(const std::string& name, glm::ivec3& vec, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool;
-        ARCADIA_API auto DragIVec3(const std::string& name, std::function<glm::ivec3()> getter, std::function<void(const glm::ivec3&)> setter, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool;
+        ARCADIA_API auto DragIVec3(const std::string& name, glm::ivec3& vec, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool;
+        ARCADIA_API auto DragIVec3(const std::string& name, std::function<glm::ivec3()> getter, std::function<void(const glm::ivec3&)> setter, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool;
 
-        ARCADIA_API auto DragIVec4(const std::string& name, glm::ivec4& vec, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool;
-        ARCADIA_API auto DragIVec4(const std::string& name, std::function<glm::ivec4()> getter, std::function<void(const glm::ivec4&)> setter, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool;
+        ARCADIA_API auto DragIVec4(const std::string& name, glm::ivec4& vec, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool;
+        ARCADIA_API auto DragIVec4(const std::string& name, std::function<glm::ivec4()> getter, std::function<void(const glm::ivec4&)> setter, float speed, int min, int max, const char* format, ImGuiSliderFlags flags) -> bool;
 
         ARCADIA_API auto DragFloat(const std::string& name, float& f, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool;
         ARCADIA_API auto DragFloat(const std::string& name, std::function<float()> getter, std::function<void(float)> setter, float speed, float min, float max, const char* format, ImGuiSliderFlags flags) -> bool;
